使用jsonwebtoken做登录

+++ server/routes/Manager.js
@@ -1,9 +1,27 @@
 const express = require('express')
 const router = express.Router()
 const Manager = require('../apps/Manager')
+const auth = require('../components/auth')
+const AUTH_ERROR = { auth: 0 } // 统一登录失败
+
+// 后台登录
+router.post('/login', async(req, res) => {
+  try {
+    let r = await Manager.login(req.body)
+    let token = auth.setToken(r.data.user)
+    // console.log('插入cookie的token:', token)
+    res.cookie('Angelebaby', token);
+    res.send(r)
+  } catch (err) {
+    res.send(err)
+  }
+})
 
 // 获取数据列表
 router.get('/getList', async(req, res, next) => {
+  let token = req.cookies.Angelebaby
+  // console.log('从cookie读取到的token:', token)
+  if (!auth.checkToken(token)) return res.send({ auth: 0 })
   try {
     res.send(await Manager.getList(req.query))
   } catch (err) {
@@ -13,6 +31,7 @@ router.get('/getList', async(req, res, next) => {
 
 // 根据ID更新单条数据
 router.post('/updateDataById', async(req, res) => {
+  if (!auth.checkToken(req.cookies.Angelebaby)) return res.send({ auth: 0 })
   try {
     res.send(await Manager.updateDataById(req.body))
   } catch (err) {
@@ -20,17 +39,9 @@ router.post('/updateDataById', async(req, res) => {
   }
 })
 
-// 后台登录
-router.post('/login', async(req, res) => {
-  try {
-    res.send(await Manager.login(req.body))
-  } catch (err) {
-    res.send(err)
-  }
-})
-
 // 设置某日是否可以提交加班订餐记录
 router.post('/setSubmit', async(req, res) => {
+  if (!auth.checkToken(req.cookies.Angelebaby)) return res.send({ auth: 0 })
   try {
     res.send(await Manager.setSubmit(req.body))
   } catch (err) {
@@ -40,6 +51,7 @@ router.post('/setSubmit', async(req, res) => {
 
 // 删除某条订餐信息记录
 router.post('/deleteOrder', async(req, res) => {
+  if (!auth.checkToken(req.cookies.Angelebaby)) return res.send({ auth: 0 })
   try {
     res.send(await Manager.deleteOrder(req.body))
   } catch (err) {
